fix: address tech lead review security and reliability issues
Security improvements:
- Add request body size limit (10MB) to prevent DoS attacks
- Add JS execution timeout (5s) to prevent infinite loop DoS
- Return 413 status for oversized request bodies

Reliability improvements:
- Add DB connection validation with ping on startup
- Implement graceful shutdown with signal handling (SIGINT/SIGTERM)
- Add HTTP status code validation (100-599 range)
- Handle resource close errors with proper logging

Performance improvements:
- Add sync.Pool for goja VM reuse in value_parser and filter
- Optimize deepCopy to use recursive copy instead of JSON marshal/unmarshal
- Preserve nil slices/maps in deepCopy

Code quality:
- Consolidate config validation into ValidateTransforms()
- Extract toHTTPStatus helper for type conversion
- Change multipart close errors from DEBUG to WARN level
- Add error key support in TransformError.Error()

+++ internal/mock/json_test.go
@@ -28,7 +28,7 @@ func TestNewJSON(t *testing.T) {
 		assert.Len(t, resultSlice, 2)
 
 		row1 := resultSlice[0].(map[string]any)
-		assert.Equal(t, float64(1), row1["id"]) // JSON unmarshal converts int to float64
+		assert.Equal(t, 1, row1["id"]) // Preserves original Go type
 		assert.Equal(t, "Alice", row1["name"])
 	})
 
@@ -41,7 +41,7 @@ func TestNewJSON(t *testing.T) {
 		require.NoError(t, err)
 
 		resultMap := result.(map[string]any)
-		assert.Equal(t, float64(1), resultMap["id"])
+		assert.Equal(t, 1, resultMap["id"]) // Preserves original Go type
 		assert.Equal(t, "Alice", resultMap["name"])
 	})
 
@@ -206,7 +206,7 @@ func TestParseJSONL(t *testing.T) {
 		result, _, err := source.Data(nil, "", nil, nil)
 		require.NoError(t, err)
 
-		// Empty JSONL returns nil (via deepCopy of nil []any)
+		// Empty JSONL returns nil (nil slice preserved by deepCopy)
 		assert.Nil(t, result)
 	})
